Lombok and Autowired Constructors
