+++ src/main/java/main/controller/ApiGeneralController.java
@@ -1,10 +1,10 @@
 package main.controller;
 
-import main.dto.CalendarInfo;
-import main.dto.SettingsResponse;
-import main.dto.TagStorage;
+import main.dto.responce.CalendarInfo;
+import main.dto.responce.SettingsResponse;
+import main.dto.responce.TagStorage;
+import main.dto.responce.InitStorage;
 import main.persistence.service.PostService;
-import main.dto.InitStorage;
 import main.persistence.service.SettingsService;
 import main.persistence.service.TagService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,17 +18,19 @@ public class ApiGeneralController {
 
     private final InitStorage initStorage;
 
-    @Autowired
-    private SettingsService settingsService;
+    private final SettingsService settingsService;
+
+    private final PostService postService;
+
+    private final TagService tagService;
 
     @Autowired
-    private PostService postService;
-
-    @Autowired
-    private TagService tagService;
-
-    public ApiGeneralController(InitStorage initStorage) {
+    public ApiGeneralController(InitStorage initStorage, SettingsService settingsService,
+                                PostService postService, TagService tagService) {
         this.initStorage = initStorage;
+        this.settingsService = settingsService;
+        this.postService = postService;
+        this.tagService = tagService;
     }
 
     @GetMapping("/init")
