--- src/main/java/main/dto/AuthRequest.java
@@ -1,26 +1,0 @@
-package main.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public class AuthRequest {
-
-    @JsonProperty("e_mail")
-    private String email;
-    private String password;
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
