--- src/main/java/main/dto/AuthResponse.java
@@ -1,27 +1,0 @@
-package main.dto;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class AuthResponse {
-
-    private boolean result;
-
-    private UserDtoResponse user;
-
-    public UserDtoResponse getUser() {
-        return user;
-    }
-
-    public void setUser(UserDtoResponse user) {
-        this.user = user;
-    }
-
-    public boolean isResult() {
-        return result;
-    }
-
-    public void setResult(boolean result) {
-        this.result = result;
-    }
-}
