--- src/main/java/main/dto/CaptchaResponse.java
@@ -1,26 +1,0 @@
-package main.dto;
-
-import java.util.Base64;
-
-public class CaptchaResponse {
-
-    private String secret;
-
-    private String image;
-
-    public String getSecret() {
-        return secret;
-    }
-
-    public void setSecret(String secret) {
-        this.secret = secret;
-    }
-
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
-    }
-}
