--- src/main/java/main/dto/RegisterDto.java
@@ -1,29 +1,0 @@
-package main.dto;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-
-import java.util.HashMap;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class RegisterDto {
-
-    private Boolean result;
-
-    private HashMap<String,String> errors;
-
-    public Boolean getResult() {
-        return result;
-    }
-
-    public void setResult(Boolean result) {
-        this.result = result;
-    }
-
-    public HashMap<String, String> getErrors() {
-        return errors;
-    }
-
-    public void setErrors(HashMap<String, String> errors) {
-        this.errors = errors;
-    }
-}
