+++ src/main/java/main/dto/responce/CalendarInfo.java
@@ -1,0 +1,16 @@
+package main.dto.responce;
+
+import lombok.Data;
+
+import java.util.Date;
+import java.util.HashMap;
+import java.util.List;
+
+@Data
+public class CalendarInfo {
+
+    List<Integer> years;
+
+    HashMap<String, Integer> posts;
+    
+}
