+++ src/main/java/main/dto/responce/CaptchaResponse.java
@@ -1,0 +1,12 @@
+package main.dto.responce;
+
+import lombok.Data;
+
+@Data
+public class CaptchaResponse {
+
+    private String secret;
+
+    private String image;
+
+}
