+++ src/main/java/main/dto/responce/InitStorage.java
@@ -1,0 +1,24 @@
+package main.dto.responce;
+
+import lombok.Data;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+@Component
+@Data
+public class InitStorage {
+
+    @Value("${blog.title}")
+    private String title;
+    @Value("${blog.subtitle}")
+    private String subtitle;
+    @Value("${blog.phone}")
+    private String phone;
+    @Value("${blog.email}")
+    private String email;
+    @Value("${blog.copyright}")
+    private String copyright;
+    @Value("${blog.copyrightFrom}")
+    private String copyrightFrom;
+
+}
