+++ src/main/java/main/dto/responce/PostsInfo.java
@@ -1,0 +1,17 @@
+package main.dto.responce;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.Data;
+
+@Data
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class PostsInfo {
+
+    private int count = 0;
+
+    private List<PostDtoResponse> posts = new ArrayList<>();
+
+}
