+++ src/main/java/main/dto/responce/RegisterDto.java
@@ -1,0 +1,16 @@
+package main.dto.responce;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.Data;
+
+import java.util.HashMap;
+
+@Data
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class RegisterDto {
+
+    private Boolean result;
+
+    private HashMap<String,String> errors;
+
+}
