+++ src/main/java/main/dto/responce/TagInfo.java
@@ -1,0 +1,14 @@
+package main.dto.responce;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.Data;
+
+@Data
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class TagInfo {
+
+    private String name;
+
+    private double weight;
+
+}
