+++ src/main/java/main/dto/responce/TagStorage.java
@@ -1,21 +1,16 @@
-package main.dto;
+package main.dto.responce;
+
+import lombok.Data;
 
 import java.util.ArrayList;
 import java.util.List;
 
+@Data
 public class TagStorage {
 
     private List<TagInfo> tags = new ArrayList<>();
 
     public void addTag(TagInfo tag){
         tags.add(tag);
-    }
-
-    public List<TagInfo> getTags() {
-        return tags;
-    }
-
-    public void setTags(List<TagInfo> tags) {
-        this.tags = tags;
     }
 }
