+++ src/main/java/main/dto/responce/UserDtoResponse.java
@@ -1,0 +1,24 @@
+package main.dto.responce;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.Data;
+
+@Data
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class UserDtoResponse {
+
+    private int id;
+
+    private String name;
+
+    private String photo;
+
+    private String email;
+
+    private Boolean settings;
+
+    private Boolean moderation;
+
+    private int moderationCount;
+
+}
